Refactor: Move DTOs and repository to their own packages, update controller logic to use DTOs
