+++ src/main/java/com/portfolio/livros/dto/DadosCadastraLivro.java
@@ -1,4 +1,4 @@
-package com.portfolio.livros.model;
+package com.portfolio.livros.dto;
 
 public record DadosCadastraLivro(String titulo, String autor) {
 }
