+++ src/main/java/com/portfolio/livros/dto/DadosEditarLivro.java
@@ -1,4 +1,4 @@
-package com.portfolio.livros.model;
+package com.portfolio.livros.dto;
 
 public record DadosEditarLivro(Long id, String titulo, String autor) {
 }
