+++ src/main/java/com/portfolio/livros/repository/LivroRepository.java
@@ -1,5 +1,6 @@
-package com.portfolio.livros.model;
+package com.portfolio.livros.repository;
 
+import com.portfolio.livros.model.Livro;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface LivroRepository extends JpaRepository<Livro, Long> {
